Replace old Cesium3DTileset with new version
- Add method to create Tileset from Cesium3DTile objects, so functionality remains similar
- Update the tests with new methods and imports
- Ignore npm files in test/run-cesium directory
- Use os.path.join to create full filepath in Cesium3DTile

Relates to #9
