+++ viz_3dtiles/__init__.py
@@ -1,8 +1,7 @@
 # -*- coding: utf-8 -*-
 from .Cesium3DTile import Cesium3DTile
-from .Cesium3DTileset import Cesium3DTileset
+from .Cesium3DTileset import Tileset, Asset, Content, Tile
 from .BoundingVolume import BoundingVolume, BoundingVolumeBox, BoundingVolumeRegion
-from .Tileset import Tileset, Asset, Content, Tile
 from .TreeGenerator import *
 
 __version__ = '0.0.1'
